backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
supabase==2.10.0
google-genai==0.5.0
python-dotenv==1.0.1
